+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- `[System.MTAThread]` and `[System.STAThread]` fixer to proper Apartment attributes.
+
 ## [1.0.1] - 2018-01-25
 This is due to 1.0.0 has been mistakenly published to Nuget.org containing broken content.